requirements.txt
opencv-python
google-generativeai
python-dotenv
pillow
numpy
streamlit
